Added option `exonerate-alignment-model` to specify alignment model for Exonerate.
This has been introduced because somebody had problems with sequences
containing what looks like introns to Exonerate. The option accepts the
values 'protein2genome' and 'protein2dna' and defaults to
'protein2genome'. The difference between the two models is that
protein2dna does not model introns. If there are intron phases in the
"transcript" sequence, the resulting ORF will be shorter because
Exonerate will stop aligning at introns, but it will be corresponding
and consistent.

Forgot to update changelog