Moved notification on reftaxon skipping to `verbose`
Also improved wording on said message. It occurs when a reverse search
is evaluated and a list of reference taxa has been specified. If the
reverse search does hit a sequence that belongs to the COG in question,
it is tested whether the sequence is of a taxon that occurs in the
reftaxa list. If not, there has been no notification so far in verbose
mode (only in debug). This notification has been moved to verbose mode,
and also made clearer (hopefully).